Timer synchronisation enhanced

+++ Source/OS9exec_core/os9exec_nt.c
@@ -41,6 +41,9 @@
  *    $Locker$ (who has reserved checkout)
  *  Log:
  *    $Log$
+ *    Revision 1.12  2002/07/23 19:38:34  bfo
+ *    introduce <syCorr> variable
+ *
  *    Revision 1.11  2002/07/21 14:35:55  bfo
  *    Synchronise the currentTick with seconds of time/date
  *
@@ -226,7 +229,7 @@ ulong   my_inetaddr   = os9_long(0x7F000001); /* loopback by default: 127.0.0.1 
 /* tickCount at start of the program */
 ulong  startTick= 0;
 ulong   lastTick= 0;
-ulong     syCorr= 0;
+int       syCorr= 0;
 
 #ifdef linux
   ulong sec0;
